@@ -5,28 +5,64 @@
   "dependencies": {
     "@accordproject/cicero-core": "0.12.2",
     "axios": "^0.18.0",
+    "babel-runtime": "^6.26.0",
     "braid": "0.0.1",
     "braid-client": "^3.3.1-20181219142722288",
     "codemirror": "^5.40.2",
     "jsonrpc-lite": "^2.0.2",
     "moment-mini": "2.22.1",
     "re-resizable": "^4.9.0",
-    "react": "^16.7.0",
+    "prop-types": "^15.7.2",
+    "react": "^16.8.6",
+    "react-dom": "^16.6.0",
     "react-codemirror2": "^5.1.0",
-    "react-dom": "^16.7.0",
+    "react-filepond": "^5.0.0",
+    "react-inspector": "^2.3.0",
     "react-resizable": "^1.7.5",
-    "react-scripts": "2.1.3",
     "semantic-ui-react": "^0.82.5",
     "styled-components": "^4.2.0"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+  "devDependencies": {
+    "@babel/cli": "^7.4.3",
+    "@babel/core": "^7.4.3",
+    "@babel/plugin-proposal-class-properties": "^7.4.0",
+    "@babel/plugin-transform-runtime": "^7.4.3",
+    "@babel/polyfill": "^7.4.3",
+    "@babel/preset-env": "^7.4.3",
+    "@babel/preset-react": "^7.0.0",
+    "@clausehq/eslint-config": "^0.1.5",
+    "babel-eslint": "^10.0.1",
+    "babel-jest": "^24.7.1",
+    "babel-loader": "^8.0.5",
+    "babel-plugin-transform-runtime": "^6.23.0",
+    "clean-webpack-plugin": "^2.0.1",
+    "css-loader": "^2.1.1",
+    "enzyme": "^3.9.0",
+    "enzyme-adapter-react-16": "^1.12.1",
+    "enzyme-to-json": "^3.3.5",
+    "eslint": "^5.16.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-config-loopback": "^13.1.0",
+    "eslint-plugin-import": "^2.16.0",
+    "eslint-plugin-jest": "^22.4.1",
+    "eslint-plugin-jsx-a11y": "^6.2.1",
+    "eslint-plugin-promise": "^4.1.1",
+    "eslint-plugin-react": "^7.12.4",
+    "file-loader": "^3.0.1",
+    "html-loader": "^0.5.5",
+    "html-webpack-plugin": "^3.2.0",
+    "jest": "^24.7.1",
+    "node-sass": "^4.12.0",
+    "sass-loader": "^7.1.0",
+    "style-loader": "^0.23.1",
+    "url-loader": "^1.1.2",
+    "webpack": "^4.29.6",
+    "webpack-cli": "^3.3.0",
+    "webpack-dev-server": "^3.3.1"
   },
-  "eslintConfig": {
-    "extends": "react-app"
+  "scripts": {
+    "start": "webpack-dev-server --config webpack.config.dev.js --mode development --watch ",
+    "build": "webpack --config webpack.config.prod.js --mode production"
   },
   "browserslist": [
     ">0.2%",
